Add tests for the confirmation banners
