+++ test/controller/learningRecord/index.spec.ts
@@ -6,7 +6,8 @@ import {client} from '../../../src/lib/service/cslService/baseConfig'
 import {LearningRecordCache} from '../../../src/lib/service/cslService/cache/learningRecordCache'
 import {setCaches} from '../../../src/lib/service/cslService/cslServiceClient'
 import * as learningRecordIndex from '../../../src/ui/controllers/learning-record/index'
-import {assertTables, TableAssertion, TextContentAsserter} from '../../utils/htmlUtils'
+import {assertTables, TableAssertion} from '../../utils/htmlAssertions/assertTable'
+import {TextContentAsserter} from '../../utils/htmlUtils'
 import {getApp} from '../../utils/testApp'
 
 const getLearningPlanTableAssertion = (
