+++ test/utils/htmlUtils.ts
@@ -1,7 +1,7 @@
 import {expect} from 'chai'
 import {JSDOM} from 'jsdom'
 
-type HTML = Document | Element
+export type HTML = Document | Element
 
 export interface HtmlAssertion {
 	querySelector: string
@@ -29,7 +29,7 @@ export const classAssertion = (classes: string[], expected: ExpectedHtmlValues |
 	}
 }
 
-interface TagContentAsserter {
+export interface TagContentAsserter {
 	assert(elem: Element, exp: Chai.ExpectStatic): void
 }
 
@@ -77,18 +77,6 @@ export const titleAssertion = (expValue: string): HtmlAssertion => {
 	}
 }
 
-export const getCTALinkButtonAssertion = (expHref: string, expText: string): HtmlAssertion => {
-	return {
-		querySelector: `.button`,
-		expected: {
-			content: new TextContentAsserter(expText),
-			attributes: {
-				href: expHref,
-			},
-		},
-	}
-}
-
 export const getBackLinkAssertion = (expHref: string, expText: string): HtmlAssertion => {
 	return {
 		querySelector: `.link-back`,
@@ -110,7 +98,9 @@ export const getAssertNotificationBanner = (expTitle: string, expContent: string
 
 export const assertHtml = (html: string | HTML, assertions: HtmlAssertion[]) => {
 	const page = typeof html === 'string' ? new JSDOM(html).window.document : html
-	assertions.forEach(a => assertDoc(page, a))
+	assertions.forEach(a => {
+		assertDoc(page, a)
+	})
 }
 
 export const assertDocMultiple = (doc: HTML, assertions: HtmlAssertion[], exp?: Chai.ExpectStatic) => {
@@ -122,6 +112,7 @@ export const assertDocMultiple = (doc: HTML, assertions: HtmlAssertion[], exp?: 
 
 export const assertDoc = (doc: HTML, assertion: HtmlAssertion, exp?: Chai.ExpectStatic) => {
 	exp = exp ? exp : expect
+	exp(doc, 'HTML document is undefined').not.eql(undefined)
 	const elem = doc.querySelectorAll(assertion.querySelector)[0]
 	const expected = assertion.expected
 	if (expected === null) {
@@ -141,330 +132,5 @@ export const assertDoc = (doc: HTML, assertion: HtmlAssertion, exp?: Chai.Expect
 				`Expected HTML attribute '${key}' to equal '${expVal}' (tag: '${elem.outerHTML}')`
 			)
 		})
-	}
-}
-
-export interface CourseDetailsAssertion {
-	expType: string
-	expDuration: string | null
-	expAreasOfWork: string[] | null
-	expLocation: string | null
-	expGrades: string[] | null
-	expCost: string | null
-}
-
-export const getDetailsRowAssertion = (expHeading: string, expText: string): HtmlAssertion[] => {
-	return [
-		{
-			querySelector: 'th.lpg-related-items__th',
-			expected: {
-				content: new TextContentAsserter(expHeading),
-			},
-		},
-		{
-			querySelector: 'td.lpg-related-items__td',
-			expected: {
-				content: new TextContentAsserter(expText),
-			},
-		},
-	]
-}
-
-export interface TableAssertion {
-	heading: TextContentAsserter[]
-	rows: TextContentAsserter[][]
-}
-
-export const assertTable = (html: string, expectedTable: TableAssertion) => {
-	const assertions: HtmlAssertion[] = []
-	for (let i = 0; i < expectedTable.heading.length; i++) {
-		const expThContent = expectedTable.heading[i]
-		const expTh: HtmlAssertion = {
-			querySelector: `thead tr th:nth-of-type(${i + 1})`,
-			expected: {
-				content: expThContent,
-			},
-		}
-		assertions.push(expTh)
-	}
-	for (let i = 0; i < expectedTable.rows.length; i++) {
-		const expRow = expectedTable.rows[i]
-		for (let j = 0; j < expRow.length; j++) {
-			const expTdContent = expRow[j]
-			const expTd: HtmlAssertion = {
-				querySelector: `tbody tr:nth-of-type(${i + 1}) td:nth-of-type(${j + 1})`,
-				expected: {
-					content: expTdContent,
-				},
-			}
-			assertions.push(expTd)
-		}
-	}
-	assertHtml(html, assertions)
-}
-
-export const assertTables = (html: string, expectedTables: TableAssertion[]) => {
-	const doc = new JSDOM(html).window.document
-	const tableHtmls = doc.getElementsByTagName('table')
-	for (let i = 0; i < expectedTables.length; i++) {
-		const expectedTable = expectedTables[i]
-		const tableHtml = tableHtmls[i]
-		assertTable(tableHtml.outerHTML, expectedTable)
-	}
-}
-
-export const assertCourseDetails = (html: string, expValues: CourseDetailsAssertion) => {
-	const doc = new JSDOM(html).window.document
-	const assertions: [HtmlAssertion[]] = [getDetailsRowAssertion('Course type', expValues.expType)]
-	if (expValues.expDuration) {
-		assertions.push(getDetailsRowAssertion('Duration', expValues.expDuration))
-	}
-	if (expValues.expAreasOfWork) {
-		assertions.push(getDetailsRowAssertion('Key area', expValues.expAreasOfWork.join(', ')))
-	}
-	if (expValues.expLocation) {
-		assertions.push(getDetailsRowAssertion('Location', expValues.expLocation))
-	}
-	if (expValues.expGrades) {
-		assertions.push(getDetailsRowAssertion('Level', expValues.expGrades.join(', ')))
-	}
-	if (expValues.expCost) {
-		assertions.push(getDetailsRowAssertion('Cost', expValues.expCost))
-	}
-	const elems = doc.querySelectorAll('tr')
-	for (let i = 0; i < elems.length; i++) {
-		const elem = elems[i]
-		const assertion = assertions[i]
-		assertDocMultiple(elem, assertion)
-	}
-}
-
-export interface ModuleCardAssertion {
-	expTitle: string
-	expDescription: string
-	expOptional: boolean
-	ctaElem: HtmlAssertion
-	details?: {
-		expType: string
-		expDuration?: string
-		expState?: string | null
-		expCost?: string | null
-	}
-}
-
-export interface CourseCardAssertion {
-	expTitle: {
-		text: string
-		href: string
-	}
-	properties: {
-		type: string
-		duration: string
-		cost?: string
-		statusBadge?: string
-	}
-	moduleCount: number
-	expDescription: string
-	dueBy?: string
-	cta: {
-		primary: {
-			href: string
-			text: string
-		}
-		secondary?: {
-			text: string
-			href?: string
-		}
-	}
-	eventModule?: {
-		title: {
-			text: string
-			href: string
-		}
-		status: string
-		type: string
-		dates: string[]
-		cta?: {
-			text: string
-			href: string
-		}
-	}
-}
-
-export const getLinkCTAAssertion = (expLaunchLink: string, expModuleTitle: string): HtmlAssertion => {
-	const assertion = getCTAAssertion(expLaunchLink, 'Start', expModuleTitle)
-	assertion.expected!.attributes!.target = '_blank'
-	return assertion
-}
-
-export const getCTAAssertion = (
-	expLaunchLink: string,
-	expModuleActionText: string,
-	expModuleTitle: string
-): HtmlAssertion => {
-	return {
-		querySelector: '.discite__action-link--main',
-		expected: {
-			content: new MultipleContentAsserter([
-				new TextContainsAsserter(expModuleActionText),
-				new TextContainsAsserter(expModuleTitle),
-			]),
-			attributes: {
-				href: expLaunchLink,
-			},
-		},
-	}
-}
-
-export const assertCourseCard = (html: HTML, expValue: CourseCardAssertion) => {
-	const assertions: HtmlAssertion[] = [
-		classAssertion(['lpg-course-link'], {
-			attributes: {
-				href: expValue.expTitle.href,
-			},
-			content: new TextContentAsserter(expValue.expTitle.text),
-		}),
-		classAssertion(['lpg-course-type'], {
-			content: new TextContentAsserter(expValue.properties.type),
-		}),
-		classAssertion(['lpg-course-duration'], {
-			content: new TextContentAsserter(expValue.properties.duration),
-		}),
-		classAssertion(['discite__counter'], {
-			content: new TextContentAsserter(`This course has ${expValue.moduleCount} modules`),
-		}),
-		classAssertion(['discite__desc'], {
-			content: new TextContentAsserter(expValue.expDescription),
-		}),
-		classAssertion(['discite__action-link--main'], {
-			content: new TextContentAsserter(expValue.cta.primary.text),
-			attributes: {
-				href: expValue.cta.primary.href,
-			},
-		}),
-	]
-	if (expValue.properties.statusBadge) {
-		assertions.push(
-			classAssertion(['badge', 'badge--discite'], {
-				content: new TextContentAsserter(expValue.properties.statusBadge),
-			})
-		)
-	}
-	if (expValue.properties.cost) {
-		assertions.push(
-			classAssertion(['lpg-course-cost'], {
-				content: new TextContentAsserter(expValue.properties.cost),
-			})
-		)
-	}
-	if (expValue.dueBy) {
-		assertions.push(
-			classAssertion(['lpg-course-dueBy'], {
-				content: new TextContentAsserter(`Due by: ${expValue.dueBy}`),
-			})
-		)
-	}
-	if (expValue.cta.secondary) {
-		assertions.push(
-			classAssertion(['discite__action-link--main:nth-of-type(2)'], {
-				content: new TextContentAsserter(expValue.cta.secondary.text),
-				attributes: {
-					href: expValue.cta.secondary.href!,
-				},
-			})
-		)
-	}
-	if (expValue.eventModule) {
-		const eventModuleHtml = html.getElementsByClassName('discite__item--event')[0]
-		const eventModuleAssertions = [
-			classAssertion(['govuk-link'], {
-				content: new TextContentAsserter(expValue.eventModule.title.text),
-				attributes: {
-					href: expValue.eventModule.title.href,
-				},
-			}),
-			classAssertion(['lpg-course-type'], {
-				content: new TextContentAsserter(expValue.eventModule.type),
-			}),
-			classAssertion(['badge', 'badge--discite', 'badge--info'], {
-				content: new TextContentAsserter(expValue.eventModule.status),
-			}),
-		]
-		for (let i = 0; i < expValue.eventModule.dates.length; i++) {
-			eventModuleAssertions.push({
-				querySelector: `li.discite__property-item:nth-of-type(${i + 1})`,
-				expected: {
-					content: new TextContentAsserter(expValue.eventModule.dates[i]),
-				},
-			})
-		}
-		assertHtml(eventModuleHtml, eventModuleAssertions)
-	}
-	assertHtml(html, assertions)
-}
-
-export const assertCourseCards = (html: string, expValues: CourseCardAssertion[]) => {
-	const page = new JSDOM(html).window.document
-	const cardHtmls = page.getElementsByClassName('discite__item')
-	for (let i = 0; i < expValues.length; i++) {
-		assertCourseCard(cardHtmls[i], expValues[i])
-	}
-}
-
-export const assertModuleCard = (html: string, expValues: ModuleCardAssertion[]) => {
-	const page = new JSDOM(html).window.document
-	const cardHtmls = page.getElementsByClassName('discite__item u-clearfix discite__item--module')
-	for (let i = 0; i < expValues.length; i++) {
-		const expValue = expValues[i]
-		const assertions: HtmlAssertion[] = [
-			classAssertion(['heading', 'bold-small', 'heading--text'], {
-				content: new TextContentAsserter(expValue.expTitle),
-			}),
-			classAssertion(['discite__description'], {
-				content: new TextContentAsserter(expValue.expDescription),
-			}),
-		]
-		if (expValue.expOptional) {
-			assertions.push(
-				classAssertion(['discite__optional'], {
-					content: new TextContentAsserter('This module is optional'),
-				})
-			)
-		}
-		assertions.push(expValue.ctaElem)
-		const details = expValue.details
-		if (details) {
-			assertions.push({
-				querySelector: 'span.lpg-course-type',
-				expected: {content: new TextContentAsserter(details.expType)},
-			})
-			if (details.expCost !== undefined) {
-				if (details.expCost !== null) {
-					assertions.push({
-						querySelector: 'span.lpg-course-cost',
-						expected: {content: new TextContentAsserter(`£${details.expCost}`)},
-					})
-				} else {
-					assertions.push({querySelector: 'span.lpg-course-cost', expected: null})
-				}
-			}
-			if (details.expDuration) {
-				assertions.push({
-					querySelector: 'span.lpg-course-duration',
-					expected: {content: new TextContentAsserter(details.expDuration)},
-				})
-			}
-			if (details.expState !== undefined) {
-				if (details.expState === null) {
-					assertions.push({querySelector: 'div.discite__status', expected: null})
-				} else {
-					assertions.push({
-						querySelector: 'span.badge.badge--info',
-						expected: {content: new TextContentAsserter(details.expState)},
-					})
-				}
-			}
-			assertHtml(cardHtmls[i].outerHTML, assertions)
-		}
 	}
 }
